+++ src/main/java/model/node/LeafNode.java
@@ -4,12 +4,14 @@ import lombok.EqualsAndHashCode;
 import lombok.Value;
 import org.jetbrains.annotations.NotNull;
 
+import java.math.BigInteger;
+
 @EqualsAndHashCode(callSuper = true)
 @Value
 public class LeafNode<V, H> extends Node<H> {
     @NotNull V value;
 
-    public LeafNode(int key, @NotNull V value, @NotNull H hash) {
+    public LeafNode(BigInteger key, @NotNull V value, @NotNull H hash) {
         super(key, hash);
         this.value = value;
     }
